requirements.txt
cffi==1.15.1
clr-loader==0.2.6
decorator==5.1.1
paho-mqtt==1.6.1
pycparser==2.21
pythonnet==3.0.3
pyuac==0.0.3
pywin32==306
PyYAML==6.0.1
tee==0.0.3
Win32Security==2.1.0